refactor(ops): rename add-chain to ops (#725)
* rename add-chain dir to scr-ops

* update justfile

* update doc

* more justfile updates

* update tests

* update module name

* update more docs

* update code comments

* update ci pipeline

* update go.work

* fix justfile typo

* rename to ops

* update tests

* update addnewchain test method name

* Update justfile

---------
