+++ docs/ops.md
@@ -1,4 +1,6 @@
-# Adding a Chain
+# Superchain Registry Operations (ops)
+
+## Adding a Chain
 
 The following are the steps you need to take to add a chain to the registry:
 
@@ -9,6 +11,7 @@ The following are the steps you need to take to add a chain to the registry:
 > added to the registry.
 
 ### 0. Fork this repository
+
 You will be raising a Pull Request from your fork to the upstream repo.
 
 We recommend only adding one chain at a time, and starting with a fresh branch of this repo for every chain.
@@ -46,7 +49,9 @@ just add-chain
 The remaining steps should then be followed to merge the config data into the registry -- a prerequisite for [promoting the chain](#promote-a-chain-to-standard) to a standard chain.
 
 ### 4. Understand output
+
 The tool will write the following data:
+
 - The main configuration source, with genesis data, and address of onchain system configuration. These are written to `superchain/configs/<superchain-target>/<chain-short-name>.toml`.
 - Hardfork override times, where they have been set, will be included. If and when a chain becomes a standard chain, a `superchain_time` is set in the chain config. From that time on, future hardfork activation times which are missing from the chain config will be inherited from superchain-wide values in the neighboring `superchain.toml` file.
 - Genesis system config data
@@ -56,13 +61,15 @@ The genesis largely consists of contracts common with other chains:
 all contract bytecode is deduplicated and hosted in the `extra/bytecodes` directory.
 
 The format is a gzipped JSON `genesis.json` file, with either:
+
 - a `alloc` attribute, structured like a standard `genesis.json`, but with `codeHash` (bytes32, `keccak256` hash of contract code) attribute per account, instead of the `code` attribute seen in standard Ethereum genesis definitions.
 - a `stateHash` attribute: to omit a large state (e.g. for networks with a re-genesis or migration history). Nodes can load the genesis block header, and state-sync to complete the node initialization.
 
 ### 5. Run tests locally
 
 Run the following commands to run the Go validation checks, for only the chain you added (replace the `<chain-id>` accordingly):
-```
+
+```shell
 just validate <chain-id>
 just validate-genesis-allocs <chain-id>
 ```
@@ -78,7 +85,7 @@ The [`validation_test.go`](./validation/validation_test.go) test declaration fil
 
 This tool will add your chain to the `chainList.toml` and `addresses.json` files, which contain summaries of all chains in the registry.
 
-```
+```shell
 just codegen
 ```
 
@@ -88,23 +95,27 @@ just codegen
 > If anything looks incorrect, please get in touch.
 
 ### 7. Open Your Pull Request
+
 When opening a PR:
+
 - Open it from a non-protected branch in your fork (e.g. avoid the `main` branch). This allows maintainers to push to your branch if needed, which streamlines the review and merge process.
 - Open one PR per chain you would like to add. This ensures the merge of one chain is not blocked by unexpected issues.
 - Once the PR is opened, please check the box to [allow edits from maintainers](https://docs.github.com/en/pull-requests/collaborating-with-pull-requests/working-with-forks/allowing-changes-to-a-pull-request-branch-created-from-a-fork).
 
 Once the PR is opened, the same automated checks you've run locally will then run on your PR, and your PR will be reviewed in due course. Once these checks pass, the PR will be merged.
 
-
 ## Promote a chain to standard
+
 This process is only possible for chains already in the registry.
 
 Run this command (replace the `<chain-id>` accordingly):
-```
+
+```shell
 just promote-to-standard <chain-id>
 ```
 
 This command will:
-* declare the chain as a standard chain
-* set the `superchain_time`, so that the chain receives future hardforks with the rest of the superchain (baked into downstream OPStack software, selected with [network flags](https://docs.optimism.io/builders/node-operators/configuration/base-config#initialization-via-network-flags)).
-* activate the full suite of validation checks for standard chains, including checks on the `ProxyAdminOwner`
+
+- declare the chain as a standard chain
+- set the `superchain_time`, so that the chain receives future hardforks with the rest of the superchain (baked into downstream OPStack software, selected with [network flags](https://docs.optimism.io/builders/node-operators/configuration/base-config#initialization-via-network-flags)).
+- activate the full suite of validation checks for standard chains, including checks on the `ProxyAdminOwner`
