+++ justfile
@@ -4,41 +4,41 @@ alias l := lint-all
 
 # Adding a chain
 add-chain:
-	go run ./add-chain add-new-chain
-	go run ./add-chain check-rollup-config
-	go run ./add-chain compress-genesis
-	go run ./add-chain check-genesis
+	go run ./ops add-new-chain
+	go run ./ops check-rollup-config
+	go run ./ops compress-genesis
+	go run ./ops check-genesis
 	just codegen
 
 # Promote a standard candidate chain to a standard chain, example: just promote-to-standard 10
 promote-to-standard CHAIN:
-	go run ./add-chain promote-to-standard --chain-id={{CHAIN}}
+	go run ./ops promote-to-standard --chain-id={{CHAIN}}
 	just codegen
 
 # Generate auto-generated files
-codegen: clean-add-chain
+codegen: clean-ops
 	CODEGEN=true go run superchain/internal/codegen/main.go
 
 # Lint all Go code
 lint-all:
-	golangci-lint run superchain/... validation/... add-chain/... --fix
+	golangci-lint run superchain/... validation/... ops/... --fix
 
 # Test all Go code
-test-all: test-add-chain test-superchain test-validation
+test-all: test-ops test-superchain test-validation
 
-# Test all Go code in the add-chain module
-test-add-chain:
+# Test all Go code in the ops module
+test-ops:
 	# We separate the first test from the rest because it generates artifacts
 	# Which need to exist before the remaining tests run.
-	TEST_DIRECTORY=./add-chain go run gotest.tools/gotestsum@latest --format testname -- -count=1 -run TestAddChain_Main
-	TEST_DIRECTORY=./add-chain/... go run gotest.tools/gotestsum@latest --format testname -- -count=1 -run '[^TestAddChain_Main]'
+	TEST_DIRECTORY=./ops go run gotest.tools/gotestsum@latest --format testname -- -count=1 -run TestOps_AddNewChain
+	TEST_DIRECTORY=./ops/... go run gotest.tools/gotestsum@latest --format testname -- -count=1 -run '[^TestOps_AddNewChain]'
 
 # Test all Go code in the superchain module
-test-superchain: clean-add-chain
+test-superchain: clean-ops
 	TEST_DIRECTORY=./superchain go run gotest.tools/gotestsum@latest --format testname
 
 # Unit test all Go code in the validation module, and do not run validation checks themselves
-test-validation: clean-add-chain
+test-validation: clean-ops
 	TEST_DIRECTORY=./validation/... go run gotest.tools/gotestsum@latest --format testname -- -run '[^TestValidation|^TestPromotion|^TestGenesisAllocs]'
 
 # Runs validation checks for any chain whose config changed
@@ -65,18 +65,18 @@ promotion-test:
 validate-standard-candidate CHAIN_ID:
 	TEST_DIRECTORY=./validation go run gotest.tools/gotestsum@latest --format testname -- -run='Promotion/.+\({{CHAIN_ID}}\)$' -count=1
 
-# Clean test files generated by the add-chain tooling
-clean-add-chain:
+# Clean test files generated by the ops tooling
+clean-ops:
 	rm -f superchain/configs/sepolia/testchain_*.toml
 	rm -f superchain/extra/sepolia/testchain_*.json.gz
 	rm -rf -- validation/genesis/validation-inputs/*-test/
 
 # Tidying all go.mod files
-tidy-all: tidy-add-chain tidy-superchain tidy-validation
+tidy-all: tidy-ops tidy-superchain tidy-validation
 
-# Tidy the add-chain go.mod file
-tidy-add-chain:
-	cd add-chain && go mod tidy
+# Tidy the ops go.mod file
+tidy-ops:
+	cd ops && go mod tidy
 
 # Tidy the superchain go.mod file
 tidy-superchain:
