+++ ops/.gitignore
