+++ ops/cmd/add-new-chain.go
@@ -13,9 +13,9 @@ import (
 	"github.com/BurntSushi/toml"
 	"github.com/ethereum-optimism/optimism/op-e2e/bindings"
 	"github.com/ethereum-optimism/optimism/op-service/retry"
-	"github.com/ethereum-optimism/superchain-registry/add-chain/config"
-	"github.com/ethereum-optimism/superchain-registry/add-chain/flags"
-	"github.com/ethereum-optimism/superchain-registry/add-chain/utils"
+	"github.com/ethereum-optimism/superchain-registry/ops/config"
+	"github.com/ethereum-optimism/superchain-registry/ops/flags"
+	"github.com/ethereum-optimism/superchain-registry/ops/utils"
 	"github.com/ethereum-optimism/superchain-registry/superchain"
 	"github.com/ethereum-optimism/superchain-registry/validation"
 	"github.com/ethereum-optimism/superchain-registry/validation/genesis"
@@ -271,7 +271,7 @@ func writeGenesisValidationMetadata(commit string, targetDir string) error {
 	// Define default metadata params:
 	// These may not be sufficient to make the genesis validation work,
 	// but we address that with some manual trial-and-error intervention
-	// involving OPLabs engineers after the add-chain command runs.
+	// involving OPLabs engineers after the add-new-chain command runs.
 	const defaultNodeVersion = "18.12.1"
 	const defaultMonorepoBuildCommand = "pnpm"
 	const defaultGenesisCreationCommand = "forge1" // See validation/genesis/commands.go
