+++ ops/go.mod
@@ -1,4 +1,4 @@
-module github.com/ethereum-optimism/superchain-registry/add-chain
+module github.com/ethereum-optimism/superchain-registry/ops
 
 go 1.22.0
